pyproject.toml
[project]
name = "llm-toolcall-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.12.1",
    "openai>=2.26.0",
    "python-dotenv==1.1.0",
]
